move tests folder into source folder
