+++ src/__mocks__/getLC.js
@@ -2,7 +2,6 @@ const setTasksLocalStorage = (task) => {
   const mockTasksContainer = [];
   mockTasksContainer.push(...task);
   const result = mockTasksContainer;
-  console.log('helllo');
   return result;
 };
